backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2024.12.13
openai-whisper==20231117
torch==2.1.1
python-multipart==0.0.6
aiofiles==23.2.1
ffmpeg-python==0.2.0
pydantic==2.5.0
cors==1.0.1 